+++ src/minecraft/mekanism/common/TileEntityBasicMachine.java
@@ -69,7 +69,7 @@ public abstract class TileEntityBasicMachine extends TileEntityElectricBlock imp
 	 * @param ticksRequired - how many ticks it takes to run a cycle
 	 * @param maxEnergy - how much energy this machine can store
 	 */
-	public TileEntityBasicMachine(String soundPath, String name, String path, int perTick, int ticksRequired, int maxEnergy)
+	public TileEntityBasicMachine(String soundPath, String name, String path, int perTick, int ticksRequired, double maxEnergy)
 	{
 		super(name, maxEnergy);
 		ENERGY_PER_TICK = perTick;
